refactor: changed directory of agents
