+++ core/src/main/java/io/deeplay/ai_agent/ExpectiminimaxAgent.java
@@ -1,4 +1,4 @@
-package io.deeplay.minimax;
+package io.deeplay.ai_agent;
 
 import io.deeplay.domain.Color;
 import io.deeplay.engine.GameState;
