+++ core/src/main/java/io/deeplay/ai_agent/MinimaxAgent.java
@@ -1,20 +1,14 @@
-package io.deeplay.minimax;
+package io.deeplay.ai_agent;
 
 import io.deeplay.domain.Color;
-import io.deeplay.domain.MoveType;
-import io.deeplay.domain.SwitchPieceType;
 import io.deeplay.engine.GameState;
 import io.deeplay.model.Board;
 import io.deeplay.model.Coordinates;
 import io.deeplay.model.move.Move;
 import io.deeplay.model.piece.*;
-import io.deeplay.model.player.Player;
 import io.deeplay.service.BoardUtil;
 
 import java.util.*;
-
-import static io.deeplay.model.player.Player.getPiecesPossibleToMove;
-import static io.deeplay.model.player.Player.getType;
 
 public class MinimaxAgent extends AbstractAiAgent {
     private Color maximizingColor;
